+++ lib/Models/Terria.ts
@@ -1705,10 +1705,13 @@ async function interpretHash(
   }
 }
 
+const containsStory = (initSource: any) =>
+  Array.isArray(initSource.stories) && initSource.stories.length;
+
 function interpretStartData(terria: Terria, startData: any, name: string) {
   // TODO: version check, filtering, etc.
 
-  if (startData.initSources) {
+  if (Array.isArray(startData.initSources)) {
     runInAction(() => {
       terria.initSources.push(
         ...startData.initSources.map((initSource: any) => {
@@ -1718,37 +1721,12 @@ function interpretStartData(terria: Terria, startData: any, name: string) {
           };
         })
       );
+
+      if (startData.initSources.some(containsStory)) {
+        terria.configParameters.showWelcomeMessage = false;
+      }
     });
   }
-
-  // if (defined(startData.version) && startData.version !== latestStartVersion) {
-  //   adjustForBackwardCompatibility(startData);
-  // }
-
-  // if (defined(terria.filterStartDataCallback)) {
-  //   startData = terria.filterStartDataCallback(startData) || startData;
-  // }
-
-  // // Include any initSources specified in the URL.
-  // if (defined(startData.initSources)) {
-  //   for (var i = 0; i < startData.initSources.length; ++i) {
-  //     var initSource = startData.initSources[i];
-  //     // avoid loading terria.json twice
-  //     if (
-  //       temporaryInitSources.indexOf(initSource) < 0 &&
-  //       !initFragmentExists(temporaryInitSources, initSource)
-  //     ) {
-  //       temporaryInitSources.push(initSource);
-  //       // Only add external files to the application's list of init sources.
-  //       if (
-  //         typeof initSource === "string" &&
-  //         persistentInitSources.indexOf(initSource) < 0
-  //       ) {
-  //         persistentInitSources.push(initSource);
-  //       }
-  //     }
-  //   }
-  // }
 }
 
 function setCustomRequestSchedulerDomainLimits(