+++ lib/ReactViews/Mobile/MobileMenu.jsx
@@ -66,9 +66,11 @@ const MobileMenu = observer(
     },
 
     runStories() {
-      this.props.viewState.storyBuilderShown = false;
-      this.props.viewState.storyShown = true;
-      this.props.viewState.mobileMenuVisible = false;
+      runInAction(() => {
+        this.props.viewState.storyBuilderShown = false;
+        this.props.viewState.storyShown = true;
+        this.props.viewState.mobileMenuVisible = false;
+      });
     },
 
     dismissSatelliteGuidanceAction() {